🏗️ (client): Move folders under src folder
