+++ apps/client/src/core/ViewController.ts
