+++ apps/client/src/core/ViewModel.ts
