+++ apps/client/src/css/additional-styles/range-slider.scss
