+++ apps/client/src/css/additional-styles/toggle-switch.scss
