+++ apps/client/src/css/additional-styles/utility-patterns.scss
