+++ apps/client/src/images/favicon.png
