+++ apps/client/src/images/hero-image.png
