+++ apps/client/src/partials/FeaturesBlocks.js
