+++ apps/client/src/partials/Footer.js
