+++ apps/client/src/partials/Newsletter.js
