+++ apps/client/src/utils/Dropdown.js
@@ -1,12 +1,8 @@
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
-import Transition from '../utils/Transition.js';
+import Transition from './Transition.js';
 
-function Dropdown({
-  children,
-  title
-}) {
-
+function Dropdown({ children, title }) {
   const [dropdownOpen, setDropdownOpen] = useState(false);
 
   return (
@@ -21,10 +17,14 @@ function Dropdown({
         className="text-gray-600 hover:text-gray-900 px-3 lg:px-5 py-2 flex items-center transition duration-150 ease-in-out"
         href="#0"
         aria-expanded={dropdownOpen}
-        onClick={(e) => e.preventDefault()}
+        onClick={e => e.preventDefault()}
       >
         {title}
-        <svg className="w-3 h-3 fill-current text-gray-500 cursor-pointer ml-1 flex-shrink-0" viewBox="0 0 12 12" xmlns="http://www.w3.org/2000/svg">
+        <svg
+          className="w-3 h-3 fill-current text-gray-500 cursor-pointer ml-1 flex-shrink-0"
+          viewBox="0 0 12 12"
+          xmlns="http://www.w3.org/2000/svg"
+        >
           <path d="M10.28 4.305L5.989 8.598 1.695 4.305A1 1 0 00.28 5.72l5 5a1 1 0 001.414 0l5-5a1 1 0 10-1.414-1.414z" />
         </svg>
       </a>
@@ -48,9 +48,6 @@ function Dropdown({
 export default Dropdown;
 
 Dropdown.propTypes = {
-  children: PropTypes.oneOfType([
-    PropTypes.arrayOf(PropTypes.element),
-    PropTypes.element.isRequired
-  ]),
+  children: PropTypes.oneOfType([PropTypes.arrayOf(PropTypes.element), PropTypes.element.isRequired]),
   title: PropTypes.string.isRequired,
 };
