+++ apps/client/src/utils/Modal.js
@@ -1,15 +1,8 @@
 import React, { useEffect, useRef } from 'react';
 import PropTypes from 'prop-types';
-import Transition from '../utils/Transition.js';
+import Transition from './Transition.js';
 
-function Modal({
-  children,
-  id,
-  ariaLabel,
-  show,
-  handleClose
-}) {
-
+function Modal({ children, id, ariaLabel, show, handleClose }) {
   const modalContent = useRef(null);
 
   // close the modal on click outside
@@ -20,7 +13,7 @@ function Modal({
     };
     document.addEventListener('click', clickHandler);
     return () => document.removeEventListener('click', clickHandler);
-  });  
+  });
 
   // close the modal if the esc key is pressed
   useEffect(() => {
@@ -31,7 +24,7 @@ function Modal({
     document.addEventListener('keydown', keyHandler);
 
     return () => document.removeEventListener('keydown', keyHandler);
-  });  
+  });
 
   return (
     <>
@@ -63,10 +56,10 @@ function Modal({
         leaveStart="opacity-100 scale-100"
         leaveEnd="opacity-0 scale-95"
       >
-        <div className="bg-white overflow-auto max-w-6xl w-full max-h-full" ref={modalContent}>          
+        <div className="bg-white overflow-auto max-w-6xl w-full max-h-full" ref={modalContent}>
           {children}
         </div>
-      </Transition>    
+      </Transition>
     </>
   );
 }
@@ -74,12 +67,9 @@ function Modal({
 export default Modal;
 
 Modal.propTypes = {
-  children: PropTypes.oneOfType([
-    PropTypes.arrayOf(PropTypes.element),
-    PropTypes.element.isRequired
-  ]),
+  children: PropTypes.oneOfType([PropTypes.arrayOf(PropTypes.element), PropTypes.element.isRequired]),
   id: PropTypes.string,
   ariaLabel: PropTypes.string,
   show: PropTypes.bool.isRequired,
-  handleClose: PropTypes.func.isRequired
+  handleClose: PropTypes.func.isRequired,
 };
